requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.22
celery==3.1.20
cffi==1.4.2
cryptography==1.1.2
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.9
django-celery==3.1.17
django-extensions==1.6.1
django-simple-history==1.8.0
gunicorn==19.4.1
httplib2==0.9.2
idna==2.0
kombu==3.0.33
oauth2==1.9.0.post1
oauthlib==1.0.3
psycopg2==2.6.1
pyasn1==0.1.9
pycparser==2.14
pyOpenSSL==0.15.1
python-openid==2.2.5
pytz==2015.7
redis==2.10.5
requests==2.9.0
requests-oauthlib==0.6.0
six==1.10.0
static3==0.6.1
Werkzeug==0.11.3
